# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SAP API",
    "version": "1.0.0",
    "description": "Kontrak rancangan SAP untuk frontend/backend. Kebutuhan otorisasi dan domain ada di API_SPEC.md. Belum merupakan layanan berjalan."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/assistant/chat": {
      "post": {
        "operationId": "assistantChat",
        "summary": "Kirim satu pesan ke SAPA dan terima jawaban + saran halaman",
        "tags": [
          "assistant"
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Nilai dari getCsrf; cocok cookie/token terverifikasi."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantChatInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantChatResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "401": {
            "description": "Authentication required (AUTH_REQUIRED / SESSION_EXPIRED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "403": {
            "description": "CSRF invalid atau SAPA dimatikan pengguna (ASSISTANT_DISABLED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "404": {
            "description": "Conversation not found (CONVERSATION_NOT_FOUND)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "422": {
            "description": "Pesan kosong/terlalu panjang atau pageContext tidak dikenal (ASSISTANT_MESSAGE_INVALID)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "429": {
            "description": "Rate limited (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                },
                "description": "Detik sebelum retry."
              },
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "503": {
            "description": "Feature flag mati atau provider tidak tersedia (ASSISTANT_UNAVAILABLE)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          }
        },
        "x-required-role": "user"
      }
    },
    "/assistant/conversations/{conversationId}": {
      "delete": {
        "operationId": "deleteAssistantConversation",
        "summary": "Hapus konteks percakapan sementara milik pengguna",
        "tags": [
          "assistant"
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "parameters": [
          {
            "name": "conversationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID opaque percakapan yang dikeluarkan server."
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Nilai dari getCsrf; cocok cookie/token terverifikasi."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AckResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "403": {
            "description": "Forbidden or CSRF invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "404": {
            "description": "Conversation not found atau bukan milik pengguna (CONVERSATION_NOT_FOUND)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                },
                "description": "Detik sebelum retry."
              },
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          },
          "503": {
            "description": "Dependency unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "headers": {
              "X-Contract-Version": {
                "$ref": "#/components/headers/ContractVersion"
              }
            }
          }
        },
        "x-required-role": "user"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Meta": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "requestId"
        ],
        "additionalProperties": false
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "code",
          "message"
        ],
        "additionalProperties": false
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "error",
          "meta"
        ],
        "additionalProperties": false
      },
      "Ack": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "additionalProperties": false
      },
...
      "PageContext": {
        "type": "string",
        "description": "Halaman aplikasi tempat pesan SAPA dikirim; hanya untuk memilih saran yang relevan. Bukan identitas/otorisasi.",
        "enum": [
          "dashboard",
          "scan",
          "my_reports",
          "areas",
          "scan_history",
          "achievements",
          "settings",
          "help"
        ]
      },
      "AssistantTarget": {
        "type": "string",
        "description": "Rute internal yang disetujui untuk tombol saran SAPA. Frontend memetakan nilai ini ke route; bukan URL bebas.",
        "enum": [
          "dashboard",
          "scan",
          "my_reports",
          "areas",
          "scan_history",
          "achievements",
          "settings",
          "help"
        ]
      },
      "SuggestedAction": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40
          },
          "target": {
            "$ref": "#/components/schemas/AssistantTarget"
          }
        },
        "required": [
          "label",
          "target"
        ],
        "additionalProperties": false
      },
...
      "AssistantChatInput": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "pageContext": {
            "$ref": "#/components/schemas/PageContext"
          },
          "conversationId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Null saat memulai percakapan; setelah itu pakai ID opaque dari server."
          }
        },
        "required": [
          "message",
          "pageContext"
        ],
        "additionalProperties": false
      },
      "AssistantChat": {
        "type": "object",
        "properties": {
          "conversationId": {
            "type": "string",
            "format": "uuid"
          },
          "reply": {
            "type": "string"
          },
          "suggestedActions": {
            "type": "array",
            "maxItems": 3,
            "items": {
              "$ref": "#/components/schemas/SuggestedAction"
            }
          }
        },
        "required": [
          "conversationId",
          "reply",
          "suggestedActions"
        ],
        "additionalProperties": false
      },
...
      "AckResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Ack"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
...
      "AssistantChatResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AssistantChat"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sap_session"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "deletionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sap_deletion"
      }
    },
    "headers": {
      "ContractVersion": {
        "description": "Contract version echoed on every response (success and error) so clients can detect contract drift.",
        "required": true,
        "schema": {
          "type": "string",
          "enum": [
            "1.0.0"
          ]
        }
      },
...
    }
  }
}